+++ R/replace_na.R
@@ -1,29 +1,33 @@
-#' Replace missing values
+#' Replace NAs with specified values
 #'
 #' @param data A data frame or vector.
-#' @param replace If `data` is a data frame, a named list giving the value to
-#'   replace `NA` with for each column. If `data` is a vector, a single value
-#'   used for replacement.
+#' @param replace If `data` is a data frame, `replace` takes a list of values,
+#'   with one value for each column that has `NA` values to be replaced.
+#'
+#'   If `data` is a vector, `replace` takes a single value. This single value
+#'   replaces all of the `NA` values in the vector.
 #' @param ... Additional arguments for methods. Currently unused.
-#' @return If `data` is a data frame, returns a data frame. If `data` is a
-#'   vector, returns a vector of class determined by the union of `data` and
-#'   `replace`.
-#' @seealso \code{\link[dplyr]{na_if}} to replace specified values with a `NA`.
-#'  \code{\link[dplyr]{coalesce}} to replace missing values with a specified
-#'   value.
-#' \code{\link[dplyr]{recode}} to more generally replace values.
+#' @return
+#' * If `data` is a data frame, `replace_na()` returns a data frame.
+#' * If `data` is a vector, `replace_na()` returns a vector, with class
+#'   given by the union of `data` and `replace`.
+#' @seealso [dplyr::na_if()] to replace specified values with `NA`s;
+#'   [dplyr::coalesce()] to replaces `NA`s with values from other vectors.
 #' @export
 #' @examples
-#' library(dplyr)
-#' df <- tibble(x = c(1, 2, NA), y = c("a", NA, "b"), z = list(1:5, NULL, 10:20))
+#' # Replace NAs in a data frame
+#' df <- tibble(x = c(1, 2, NA), y = c("a", NA, "b"))
 #' df %>% replace_na(list(x = 0, y = "unknown"))
-#' df %>% mutate(x = replace_na(x, 0))
 #'
-#' # NULL are the list-col equivalent of NAs
-#' df %>% replace_na(list(z = list(5)))
-#'
+#' # Replace NAs in a vector
+#' df %>% dplyr::mutate(x = replace_na(x, 0))
+#' # OR
 #' df$x %>% replace_na(0)
 #' df$y %>% replace_na("unknown")
+#'
+#' # Replace NULLs in a list: NULLs are the list-col equivalent of NAs
+#' df_list <- tibble(z = list(1:5, NULL, 10:20))
+#' df_list %>% replace_na(list(z = list(5)))
 replace_na <- function(data, replace, ...) {
   ellipsis::check_dots_used()
   UseMethod("replace_na")